testes passando

+++ src/main/java/david/augusto/luan/Calcular.java
@@ -6,18 +6,21 @@ import java.util.function.UnaryOperator;
 
 public class Calcular {
 
+    // funcional
     public int calcularValor() {
         UnaryOperator<Integer> valorCalcular = valor -> valor * 3; // paradigma funcional
         int valor = 10;
         return valorCalcular.apply(valor);
     }
 
-    public double calcularMedia() {
-        UnaryOperator<Double> valorCalcular = valor -> (valor + valor + valor) / 3;
-        double valor = 10;
+    // funcional
+    public int calcularMedia() {
+        UnaryOperator<Integer> valorCalcular = valor -> (valor + valor + valor) / 3;
+        int valor = 10;
         return valorCalcular.apply(valor);
     }
 
+    // imperativo
     public int calcularComposicaoDeFuncoes() {
         int[] valores = {1, 2, 3, 4, 5};
         AtomicInteger valor = new AtomicInteger();
